+++ Runtime/AuthenticationModel.cs
@@ -8,9 +8,25 @@ using com.noctuagames.sdk.UI;
 using Cysharp.Threading.Tasks;
 using UnityEngine;
 using UnityEngine.UIElements;
+using UnityEngine.Scripting;
+using System;
+using System.Text;
+using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
+using Newtonsoft.Json.Serialization;
+
 
 namespace com.noctuagames.sdk
 {
+
+        [Preserve]
+        public class PendingPurchaseItem
+        {
+            public int OrderId;
+            public string Timestamp;
+            public string OrderRequest;
+            public string VerifyOrderRequest;
+        }
     /*
     We were using Model-View-Presenter, further reading:
     - https://en.wikipedia.org/wiki/Model%E2%80%93view%E2%80%93presenter
@@ -53,9 +69,12 @@ namespace com.noctuagames.sdk
         private readonly AccountDeletionConfirmationDialogPresenter _accountDeletionConfirmationDialog;
         private readonly BindConfirmationDialogPresenter _bindConfirmation;
         private readonly ConnectConflictDialogPresenter _connectConflictDialog;
+        private readonly PendingPurchasesDialogPresenter _pendingPurchasesDialog;
+        private List<PendingPurchaseItem> _pendingPurchases = new List<PendingPurchaseItem>();
         private readonly WelcomeNotificationPresenter _welcome;
 
         private NoctuaAuthenticationService _authService;
+        public NoctuaIAPService _iapService;
         private GameObject _socialAuthObject;
         private SocialAuthenticationService _socialAuth;
         
@@ -67,18 +86,35 @@ namespace com.noctuagames.sdk
 
         private AuthType _currentAuthType = AuthType.Switch;
 
+        private readonly NoctuaLocale _locale;
+
+        private readonly ILogger _log = new NoctuaLogger(typeof(AuthenticationModel));
+
         internal AuthenticationModel(
             UIFactory uiFactory, 
             NoctuaAuthenticationService authService, 
+            NoctuaIAPService iapService,
             GlobalConfig config,
-            EventSender eventSender = null
+            EventSender eventSender = null,
+            NoctuaLocale locale = null
         )
         {
             _uiFactory = uiFactory;
+
+            if (locale != null)
+            {
+                _locale = locale;
+            }
+
+            if (iapService != null)
+            {
+                _iapService = iapService;
+            }
             
             _userCenter = _uiFactory.Create<UserCenterPresenter, AuthenticationModel>(this);
             _userCenter.EventSender = eventSender;
             _accountSelectionDialog = _uiFactory.Create<AccountSelectionDialogPresenter, AuthenticationModel>(this);
+            _pendingPurchasesDialog = _uiFactory.Create<PendingPurchasesDialogPresenter, AuthenticationModel>(this);
             _switchAccountConfirmationDialog = _uiFactory.Create<SwitchAccountConfirmationDialogPresenter, AuthenticationModel>(this);
             _loginOptionsDialog = _uiFactory.Create<LoginOptionsDialogPresenter, AuthenticationModel>(this);
             _emailLoginDialog = _uiFactory.Create<EmailLoginDialogPresenter, AuthenticationModel>(this);
@@ -92,6 +128,7 @@ namespace com.noctuagames.sdk
             _bindConfirmation = _uiFactory.Create<BindConfirmationDialogPresenter, AuthenticationModel>(this);
             _connectConflictDialog = _uiFactory.Create<ConnectConflictDialogPresenter, AuthenticationModel>(this);
             _welcome = _uiFactory.Create<WelcomeNotificationPresenter, AuthenticationModel>(this);
+
 
             _welcome.SetBehaviourWhitelabel(config);
             _emailLoginDialog.SetBehaviourWhitelabel(config);
@@ -255,6 +292,42 @@ namespace com.noctuagames.sdk
         public void ShowConnectConflict(PlayerToken playerToken)
         {
             _connectConflictDialog.Show(playerToken);
+        }
+
+        public string GetLanguage()
+        {
+            if (_locale == null)
+            {
+                return "en";
+            }
+
+            return _locale.GetLanguage();
+        }
+
+        private List<PendingPurchaseItem> GetPendingPurchases()
+        {
+            if (_iapService == null) {
+                return _pendingPurchases;
+            }
+
+            var list =  _iapService.GetPendingPurchases();
+            _pendingPurchases = new List<PendingPurchaseItem>();
+            foreach (var item in list)
+            {
+                _pendingPurchases.Add(new PendingPurchaseItem{
+                    OrderId = item.OrderId,
+                    Timestamp = item.OrderRequest.Timestamp,
+                    OrderRequest = Convert.ToBase64String(Encoding.UTF8.GetBytes(JsonConvert.SerializeObject(item.OrderRequest))),
+                    VerifyOrderRequest = Convert.ToBase64String(Encoding.UTF8.GetBytes(JsonConvert.SerializeObject(item.VerifyOrderRequest)))
+                });
+            }
+
+            return _pendingPurchases;
+        }
+
+        public async void ShowPendingPurchasesDialog()
+        {
+            _pendingPurchasesDialog.Show(GetPendingPurchases());
         }
     }