+++ Runtime/Http.cs
@@ -102,6 +102,13 @@ namespace com.noctuagames.sdk
             _jsonSettings.Converters.Add(new StringEnumConverter());
             _request.method = method.ToString().ToUpper();
             _request.url = url;
+
+            // Inject locale data
+            _request.SetRequestHeader("Accept-Language", Noctua.Platform.Locale.GetLanguage());
+            _request.SetRequestHeader("X-LANGUAGE", Noctua.Platform.Locale.GetLanguage());
+            _request.SetRequestHeader("X-COUNTRY", Noctua.Platform.Locale.GetCountry());
+            _request.SetRequestHeader("X-CURRENCY", Noctua.Platform.Locale.GetCurrency());
+
         }
 
         public HttpRequest WithPathParam(string key, string value)
@@ -245,7 +252,11 @@ namespace com.noctuagames.sdk
                 $"Content-Type: {_request.GetRequestHeader("Content-Type")}\n"   +
                 $"{auth}\n" +
                 $"X-CLIENT-ID: {_request.GetRequestHeader("X-CLIENT-ID")}\n"     +
-                $"X-BUNDLE-ID: {_request.GetRequestHeader("X-BUNDLE-ID")}\n\n"   +
+                $"X-BUNDLE-ID: {_request.GetRequestHeader("X-BUNDLE-ID")}\n"   +
+                $"X-LANGUAGE: {_request.GetRequestHeader("X-LANGUAGE")}\n"     +
+                $"X-COUNTRY: {_request.GetRequestHeader("X-COUNTRY")}\n"     +
+                $"X-CURRENCY: {_request.GetRequestHeader("X-CURRENCY")}\n"     +
+                $"Accept-Language: {_request.GetRequestHeader("X-LANGUAGE")}\n\n"     +
                 $"{Encoding.UTF8.GetString(_request.uploadHandler?.data ?? Array.Empty<byte>())}"
             );