+++ Runtime/Noctua.cs
@@ -311,7 +311,6 @@ namespace com.noctuagames.sdk
                 eventSender: _eventSender
             );
 
-            _auth = new NoctuaAuthentication(authService, uiFactory, config, _eventSender);
             
             var accessTokenProvider = new AccessTokenProvider(authService);
 
@@ -327,6 +326,8 @@ namespace com.noctuagames.sdk
                 _nativePlugin,
                 _eventSender
             );
+
+            _auth = new NoctuaAuthentication(authService, _iap, uiFactory, config, _eventSender, locale);
 
             _game = new NoctuaGameService(
                 new NoctuaGameService.Config
@@ -503,7 +504,7 @@ namespace com.noctuagames.sdk
             
             if (Noctua.IsFirstOpen())
             {
-                Instance.Value._eventSender.Send("first_open");
+                Instance.Value._eventSender.Send("sdk_first_open");
             }
 
             log.Info("Noctua.InitAsync() completed");